# backend/openapi_doc/specs/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Routine Cloud API v1",
    "description": "Routine Cloud API SDK",
    "version": "v1"
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User Me",
        "description": "Get current user information and sync with OIDC",
        "operationId": "users_me",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Current User Me",
        "description": "Update current user profile (profile picture)",
        "operationId": "users_update_me",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "description": "List all users.",
        "operationId": "users_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "title": "Response Users List"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List My Tasks",
        "description": "List all tasks owned by the current user.",
        "operationId": "tasks_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response Tasks List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "description": "Create a new task owned by the current user. Returns the created task.",
        "operationId": "tasks_create",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "Update Task",
        "description": "Update mutable fields of a task. All fields are optional; only provided fields are updated.",
        "operationId": "tasks_update",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete Task",
        "description": "Delete a task only if you are its owner and it is not used in routines owned by other users.",
        "operationId": "tasks_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/routines/{routine_id}/tasks": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routine Tasks",
        "operationId": "routines_tasks_list",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks List"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Add Task To Routine",
        "operationId": "routines_tasks_add",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineTaskAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks Add"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/tasks/{position}": {
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Remove Task From Routine",
        "operationId": "routines_tasks_remove",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          },
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position"
            }
          },
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskInRoutineRead"
                  },
                  "title": "Response Routines Tasks Remove"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoutineTaskAdd": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "position": {
            "title": "Position",
            "type": "integer",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "RoutineTaskAdd"
      },
...
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "sound": {
            "type": "string",
            "title": "Sound"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon_name",
          "sound",
          "duration"
        ],
        "title": "TaskCreate"
      },
      "TaskInRoutineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "sound": {
            "type": "string",
            "title": "Sound"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "icon_name",
          "sound",
          "duration",
          "position"
        ],
        "title": "TaskInRoutineRead"
      },
      "TaskRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "sound": {
            "type": "string",
            "title": "Sound"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon_name",
          "sound",
          "duration",
          "id"
        ],
        "title": "Task"
      },
      "TaskUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "nullable": true
          },
          "icon_name": {
            "title": "Icon Name",
            "type": "string",
            "nullable": true
          },
          "sound": {
            "title": "Sound",
            "type": "string",
            "nullable": true
          },
          "duration": {
            "title": "Duration",
            "type": "integer",
            "nullable": true
          }
        },
        "type": "object",
        "title": "TaskUpdate"
      },
      "UserRead": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "username": {
            "title": "Username",
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "title": "Is Active",
            "default": true,
            "type": "boolean",
            "nullable": true
          },
          "is_superuser": {
            "title": "Is Superuser",
            "default": false,
            "type": "boolean",
            "nullable": true
          },
          "oidc_sub": {
            "type": "string",
            "title": "Oidc Sub"
          },
          "friend_code": {
            "title": "Friend Code",
            "type": "string",
            "nullable": true
          },
          "profile_picture": {
            "title": "Profile Picture",
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "oidc_sub",
          "id"
        ],
        "title": "UserRead",
        "description": "Schema for returning user data."
      },
      "UserUpdate": {
        "properties": {
          "profile_picture": {
            "title": "Profile Picture",
            "type": "string",
            "nullable": true
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Schema for updating a user."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}